requirements.txt
# Core
anthropic==0.39.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2

# LangChain ecosystem
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-community==0.3.5
langchain-chroma==0.1.4

# Vector store
chromadb==0.5.23

# SQL
sqlalchemy==2.0.23

# Web search (we'll use Tavily - has free tier)
tavily-python==0.5.0

# UI
streamlit==1.39.0
plotly==5.18.0

# Embeddings
sentence-transformers==2.2.2

# Utils
openpyxl==3.1.2
kaggle==1.6.17